requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
dotenv==0.9.9
idna==3.10
python-dotenv==1.1.1
requests==2.32.4
urllib3==2.5.0
textual>=5.0.0
html2text>=2025.4.0
google-generativeai>=0.8.0
